requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
databases==0.7.0
ecdsa==0.19.0
fastapi==0.95.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.21
pydantic_core==2.27.2
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.47
starlette==0.26.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
pytest==8.3.4
pytest-asyncio==0.21.0
httpx==0.24.1